requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
tornado==6.1
tzlocal==5.2
